+++ app/docs/[slug]/page.tsx
@@ -1,699 +1,136 @@
-// "use client";
+import React from "react";
 
-import dynamic from "next/dynamic";
-import React, { useRef, useEffect } from "react";
-import Image from "next/image";
+// Import Layout and Navigation Components
+import NavBar from "../components/NavBar";
+import SidebarDocsTab from "../components/SidebarDocsTab";
+import { getFolderFiles } from "../components/SidebarDocsTab";
+
+// Import Markdown/MDX and Style tools
 import ReactMarkdown from "react-markdown";
-
-import NavBar from "../../components/NavBar";
 import styles from "./page.module.css";
-import { MDXProvider } from "@mdx-js/react";
 import matter from "gray-matter";
-// import {Button} from "flowbite-react";
-// import renderToString from 'next-mdx-remote/render-to-string'
-// import hydrate from 'next-mdx-remote/hydrate'
+
+// Import Node.js Utils
 import fs from "fs";
 import path from "path";
 
-import { CustomH1, CustomH2 } from "../components/CustomHeadings";
+// Import MDX User Components
+import { CustomH1, CustomH2 } from "../components/user/CustomHeadings";
+import { CustomImage } from "../components/user/CustomImage";
+import { CustomDivider } from "../components/user/CustomLayout";
 
-// import Content from '../pages/hello.mdx'
-
-// function convertMdxToFlowbite(fooRef : any) {
-//     const header_nodelist = fooRef.current.querySelectorAll("h1");
-//     var header_array = [...header_nodelist]; // converts NodeList to Array
-//     console.log('header_array: ', header_array)
-//     header_array.forEach((element) => {
-//         element.classList.add("helloworld")
-//         // element.className = "mb-4 text-3xl font-extrabold leading-tight text-gray-900 lg:mb-6 lg:text-4xl dark:text-white";
-//     });
-// }
-
-export default function Post({ params }: { params: { slug: string } }) {
-  if (typeof window !== "undefined") {
-  const source = fs.readFileSync(
-    path.resolve("app/docs/pages/", `${params.slug}.mdx`),
-    "utf8"
-  );
-
-  const { data, content } = matter(source);
-  const markdownBody = content;
-  }
-
-  if (typeof window !== "undefined") {
-    var themeToggleDarkIcon = document.getElementById("theme-toggle-dark-icon");
-    var themeToggleLightIcon = document.getElementById(
-      "theme-toggle-light-icon"
+export default function Doc({ params }: { params: { slug: string } }) {
+    // Get this page's markdown data
+    const source = fs.readFileSync(
+      path.resolve("app/docs/pages/", `${params.slug}.mdx`),
+      "utf8"
     );
 
-    // Change the icons inside the button based on previous settings
-    if (themeToggleLightIcon != null && themeToggleDarkIcon != null) {
-      if (
-        localStorage.getItem("color-theme") === "dark" ||
-        (!("color-theme" in localStorage) &&
-          window.matchMedia("(prefers-color-scheme: dark)").matches)
-      ) {
-        themeToggleLightIcon.style.display == "block";
-      } else {
-        themeToggleDarkIcon.style.display == "block";
-      }
+    const { data, content } = matter(source);
+    const markdownBody = content;
 
-      var themeToggleBtn = document.getElementById("theme-toggle");
+    let sidebarTabs: any = getFolderFiles(path.resolve("app/docs/pages/"), []);
+    console.log(JSON.stringify(sidebarTabs));
 
-      if (themeToggleBtn != null) {
-      themeToggleBtn.addEventListener("click", function () {
-        // toggle icons inside button
-        if (themeToggleLightIcon != null && themeToggleDarkIcon != null) {
-        console.log('computed style map: ', localStorage.getItem("color-theme") === "light")
-        themeToggleDarkIcon.style.display == ((localStorage.getItem("color-theme") === "light") ? "block" : "none") //.toggle("hidden");
-        themeToggleLightIcon.style.display == ((localStorage.getItem("color-theme") === "dark") ? "block" : "none") // .classList.toggle("hidden");
-        }
-        // if (localStorage.getItem("color-theme") === "light") {
-            
-        // }
+    // Setup Sidebar Toggle
 
-        // if set via local storage previously
-        if (localStorage.getItem("color-theme")) {
-            console.log('set previously')
-            
-          if (localStorage.getItem("color-theme") === "light") {
-            document.documentElement.classList.add("dark");
-            localStorage.setItem("color-theme", "dark");
-          } else {
-            document.documentElement.classList.remove("dark");
-            localStorage.setItem("color-theme", "light");
-          }
-
-          // if NOT set via local storage previously
-        } else {
-          if (document.documentElement.classList.contains("dark")) {
-            document.documentElement.classList.remove("dark");
-            localStorage.setItem("color-theme", "light");
-          } else {
-            document.documentElement.classList.add("dark");
-            localStorage.setItem("color-theme", "dark");
-          }
-        }
-      });
-    }
-    }
-  }
-
-  const SidebarDocsTab = ({ item }: any) => {
     return (
-      <>
-        {/* <li>
-          <a href="#" className="flex items-center p-2 text-base font-normal text-gray-900 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group">
-              <svg aria-hidden="true" className="w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M2 10a8 8 0 018-8v8h8a8 8 0 11-16 0z"></path><path d="M12 2.252A8.014 8.014 0 0117.748 8H12V2.252z"></path></svg>
-              <span className="ml-3">Overview</span>
-          </a>
-      </li> */}
-        <li key={item.id}>
-          <button
-            type="button"
-            className="flex items-center p-2 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700"
-            aria-controls="dropdown-pages"
-            data-collapse-toggle="dropdown-pages"
-          >
-            {(item.children.length == 0) ? (
-            <svg
-              aria-hidden="true"
-              className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 group-hover:text-gray-900 dark:text-gray-400 dark:group-hover:text-white"
-              fill="currentColor"
-              viewBox="0 0 20 20"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                fillRule="evenodd"
-                d="M4 4a2 2 0 012-2h4.586A2 2 0 0112 2.586L15.414 6A2 2 0 0116 7.414V16a2 2 0 01-2 2H6a2 2 0 01-2-2V4zm2 6a1 1 0 011-1h6a1 1 0 110 2H7a1 1 0 01-1-1zm1 3a1 1 0 100 2h6a1 1 0 100-2H7z"
-                clipRule="evenodd"
-              ></path>
-            </svg>
-            ) : (
-              <svg
-              aria-hidden="true"
-              className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white"
-              fill="currentColor"
-              viewBox="0 0 20 20"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path d="M7 3a1 1 0 000 2h6a1 1 0 100-2H7zM4 7a1 1 0 011-1h10a1 1 0 110 2H5a1 1 0 01-1-1zM2 11a2 2 0 012-2h12a2 2 0 012 2v4a2 2 0 01-2 2H4a2 2 0 01-2-2v-4z"></path>
-            </svg>
-            )}
-            <span className="flex-1 ml-3 text-left whitespace-nowrap">
-              {item.title}
-            </span>
-            <svg
-              aria-hidden="true"
-              className="w-6 h-6"
-              fill="currentColor"
-              viewBox="0 0 20 20"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                fillRule="evenodd"
-                d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
-                clipRule="evenodd"
-              ></path>
-            </svg>
-          </button>
-          <ul id="dropdown-pages" className={`${styles.indent} py-2 space-y-2`}>
-            {item.children.length > 0
-              ? item.children.map((item: any) => (
-                  <SidebarDocsTab key={item} item={item} />
-                ))
-              : (null
-              //   <li>
-              //   <a
-              //     href="#"
-              //     className="flex items-center p-2 pl-11 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700"
-              //   >
-              //       {item.title}
-              //   </a>
-              // </li>
-              )}
+      <div className="antialiased bg-gray-50 dark:bg-gray-900">
+        <NavBar />
 
-          </ul>
-        </li>
-        {/* <li>
-          <button type="button" className="flex items-center p-2 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700" aria-controls="dropdown-sales" data-collapse-toggle="dropdown-sales">
-              <svg aria-hidden="true" className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 group-hover:text-gray-900 dark:text-gray-400 dark:group-hover:text-white" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fillRule="evenodd" d="M10 2a4 4 0 00-4 4v1H5a1 1 0 00-.994.89l-1 9A1 1 0 004 18h12a1 1 0 00.994-1.11l-1-9A1 1 0 0015 7h-1V6a4 4 0 00-4-4zm2 5V6a2 2 0 10-4 0v1h4zm-6 3a1 1 0 112 0 1 1 0 01-2 0zm7-1a1 1 0 100 2 1 1 0 000-2z" clipRule="evenodd"></path></svg>
-              <span className="flex-1 ml-3 text-left whitespace-nowrap">Sales</span>
-              <svg aria-hidden="true" className="w-6 h-6" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fillRule="evenodd" d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z" clipRule="evenodd"></path></svg>
-          </button>
-          <ul id="dropdown-sales" className="hidden py-2 space-y-2">
-                  <li>
-                      <a href="#" className="flex items-center p-2 pl-11 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700">Products</a>
-                  </li>
-                  <li>
-                      <a href="#" className="flex items-center p-2 pl-11 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700">Billing</a>
-                  </li>
-                  <li>
-                      <a href="#" className="flex items-center p-2 pl-11 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700">Invoice</a>
-                  </li>
-          </ul>
-      </li> */}
-        {/* <li>
-          <a href="#" className="flex items-center p-2 text-base font-normal text-gray-900 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group">
-              <svg aria-hidden="true" className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M8.707 7.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l2-2a1 1 0 00-1.414-1.414L11 7.586V3a1 1 0 10-2 0v4.586l-.293-.293z"></path><path d="M3 5a2 2 0 012-2h1a1 1 0 010 2H5v7h2l1 2h4l1-2h2V5h-1a1 1 0 110-2h1a2 2 0 012 2v10a2 2 0 01-2 2H5a2 2 0 01-2-2V5z"></path></svg>
-              <span className="flex-1 ml-3 whitespace-nowrap">Messages</span>
-              <span className="inline-flex justify-center items-center w-5 h-5 text-xs font-semibold rounded-full text-primary-800 bg-primary-100 dark:bg-primary-200 dark:text-primary-800">
-                  6   
-              </span>
-          </a>
-      </li>
-      <li>
-          <button type="button" className="flex items-center p-2 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700" aria-controls="dropdown-authentication" data-collapse-toggle="dropdown-authentication">
-              <svg aria-hidden="true" className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 group-hover:text-gray-900 dark:text-gray-400 dark:group-hover:text-white" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fillRule="evenodd" d="M5 9V7a5 5 0 0110 0v2a2 2 0 012 2v5a2 2 0 01-2 2H5a2 2 0 01-2-2v-5a2 2 0 012-2zm8-2v2H7V7a3 3 0 016 0z" clipRule="evenodd"></path></svg>
-              <span className="flex-1 ml-3 text-left whitespace-nowrap">Authentication</span>
-              <svg aria-hidden="true" className="w-6 h-6" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fillRule="evenodd" d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z" clipRule="evenodd"></path></svg>
-          </button>
-          <ul id="dropdown-authentication" className="hidden py-2 space-y-2">
+        <aside
+          id="default-sidebar"
+          className={`${styles.sidebar} fixed top-0 left-0 z-40 w-64 h-screen transition-transform -translate-x-full sm:translate-x-0`}
+          aria-label="Sidenav"
+        >
+          <div className="overflow-y-auto py-5 px-3 h-full bg-white border-r border-gray-200 dark:bg-gray-800 dark:border-gray-700">
+            <ul className="space-y-2">
+              {sidebarTabs.map((item: any) => (
+                <SidebarDocsTab key={item} item={item} styles={styles} />
+              ))}
+            </ul>
+
+            <ul className="pt-5 mt-5 space-y-2 border-t border-gray-200 dark:border-gray-700">
               <li>
-                  <a href="#" className="flex items-center p-2 pl-11 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700">Sign In</a>
-              </li>
-              <li>
-                  <a href="#" className="flex items-center p-2 pl-11 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700">Sign Up</a>
-              </li>
-              <li>
-                  <a href="#" className="flex items-center p-2 pl-11 w-full text-base font-normal text-gray-900 rounded-lg transition duration-75 group hover:bg-gray-100 dark:text-white dark:hover:bg-gray-700">Forgot Password</a>
-              </li>
-          </ul>
-      </li> */}
-      </>
-    );
-  };
-
-  // const fooRef = useRef(null);
-
-  // useEffect(() => {
-  //     console.log('effect')
-  //     if (typeof window !== "undefined") {
-
-  //     convertMdxToFlowbite(fooRef)
-  //     }
-  //     }, [fooRef]);
-  //   getPostData(params.slug)
-  //     .then((Content) => {
-  //       console.log("the title is", params.slug);
-
-  //     })
-  //     .catch((err) => {
-  //       console.log(err);
-  //       return <div>ERROR</div>;
-  //     });
-  // const components = {
-  //   h1: CustomH1,
-  //   h2: CustomH2,
-
-  //   // Add other custom components for different heading levels if needed.
-  // };
-  const Component = dynamic(() => import(`../pages/${params.slug}.mdx`));
-  // console.log('front matter', matter(Component).content)
-  function getFolderFiles(folderPath: string, structure : any[]) {
-    const rawFileStructure= fs.readdirSync(
-      path.resolve(folderPath),
-      { withFileTypes: true });
-    const mutatedStrtucture = structure;
-    for (const file of rawFileStructure) {
-      if (file.isDirectory()) {
-        console.log('recursing through', file)
-        mutatedStrtucture.push({
-          id: file.name,
-          title: file.name,
-          children: getFolderFiles(path.join(folderPath, file.name), []),
-        });
-      } else {
-        const markdownWithMeta = fs.readFileSync(path.resolve('app/docs/pages', file.name), 'utf-8')
-        console.log('front matter', matter(markdownWithMeta))
-        const {data: frontMatter} = matter(markdownWithMeta)
-        console.log(frontMatter.title)
-        mutatedStrtucture.push({
-          id: file.name,
-          title: file.name,
-          children: [],
-        });
-      }
-    }
-    return mutatedStrtucture;
-    }
-    let sidebarTabs : any = getFolderFiles(path.resolve("app/docs/pages/"), [])
-    console.log(JSON.stringify(sidebarTabs))
-
-  // const sidebarTabs = [
-  //   {
-  //     id: "getting-started",
-  //     title: "Getting Started",
-  //     children: [
-  //       {
-  //         id: "getting-started",
-  //         title: "Getting Started",
-  //         children: [
-  //           {
-  //             id: "getting-started",
-  //             title: "Getting Started",
-  //             children: [],
-  //           },
-
-  //         ],
-  //       },
-  //       {
-  //         id: "getting-started",
-  //         title: "Getting Started",
-  //         children: [],
-  //       },
-  //     ],
-  //   },
-  // ];
-
-  // const sidebarTabs = docStructure;
-
-  return (
-    <div className="antialiased bg-gray-50 dark:bg-gray-900">
-      <nav className="bg-white border-b border-gray-200 px-4 py-2.5 dark:bg-gray-800 dark:border-gray-700 fixed left-0 right-0 top-0 z-50">
-        <div className="flex flex-wrap justify-between items-center">
-          <div className="flex justify-start items-center">
-            <button
-              data-drawer-target="drawer-navigation"
-              data-drawer-toggle="drawer-navigation"
-              aria-controls="drawer-navigation"
-              className="p-2 mr-2 text-gray-600 rounded-lg cursor-pointer md:hidden hover:text-gray-900 hover:bg-gray-100 focus:bg-gray-100 dark:focus:bg-gray-700 focus:ring-2 focus:ring-gray-100 dark:focus:ring-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-            >
-              <svg
-                aria-hidden="true"
-                className="w-6 h-6"
-                fill="currentColor"
-                viewBox="0 0 20 20"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  fillRule="evenodd"
-                  d="M3 5a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zM3 10a1 1 0 011-1h6a1 1 0 110 2H4a1 1 0 01-1-1zM3 15a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1z"
-                  clipRule="evenodd"
-                ></path>
-              </svg>
-              <svg
-                aria-hidden="true"
-                className="hidden w-6 h-6"
-                fill="currentColor"
-                viewBox="0 0 20 20"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  fillRule="evenodd"
-                  d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z"
-                  clipRule="evenodd"
-                ></path>
-              </svg>
-              <span className="sr-only">Toggle sidebar</span>
-            </button>
-            <a
-              href="https://wasteofplus.github.io"
-              className="flex items-center justify-between mr-4"
-            >
-              <Image
-                src="/wasteofplus.svg"
-                className="logo mr-3 h-10"
-                alt="Wasteof.plus Logo"
-                width={40}
-                height={40}
-              />
-              <span className="self-center text-2xl font-semibold whitespace-nowrap dark:text-white">
-                wasteof.plus
-              </span>
-            </a>
-            <form action="#" method="GET" className="hidden md:block md:pl-2">
-              <label className="sr-only">Search</label>
-              <div className="relative md:w-64 md:w-96">
-                <div className="flex absolute inset-y-0 left-0 items-center pl-3 pointer-events-none">
+                <a
+                  href="#"
+                  className="flex items-center p-2 text-base font-normal text-gray-900 rounded-lg transition duration-75 hover:bg-gray-100 dark:hover:bg-gray-700 dark:text-white group"
+                >
                   <svg
-                    className="w-5 h-5 text-gray-500 dark:text-gray-400"
+                    aria-hidden="true"
+                    className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white"
+                    fill="currentColor"
+                    viewBox="0 0 20 20"
+                    xmlns="http://www.w3.org/2000/svg"
+                  >
+                    <path d="M9 2a1 1 0 000 2h2a1 1 0 100-2H9z"></path>
+                    <path
+                      fillRule="evenodd"
+                      d="M4 5a2 2 0 012-2 3 3 0 003 3h2a3 3 0 003-3 2 2 0 012 2v11a2 2 0 01-2 2H6a2 2 0 01-2-2V5zm3 4a1 1 0 000 2h.01a1 1 0 100-2H7zm3 0a1 1 0 000 2h3a1 1 0 100-2h-3zm-3 4a1 1 0 100 2h.01a1 1 0 100-2H7zm3 0a1 1 0 100 2h3a1 1 0 100-2h-3z"
+                      clipRule="evenodd"
+                    ></path>
+                  </svg>
+                  <span className="ml-3">Docs</span>
+                </a>
+              </li>
+              <li>
+                <a
+                  href="#"
+                  className="flex items-center p-2 text-base font-normal text-gray-900 rounded-lg transition duration-75 hover:bg-gray-100 dark:hover:bg-gray-700 dark:text-white group"
+                >
+                  <svg
+                    aria-hidden="true"
+                    className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white"
+                    fill="currentColor"
+                    viewBox="0 0 20 20"
+                    xmlns="http://www.w3.org/2000/svg"
+                  >
+                    <path d="M7 3a1 1 0 000 2h6a1 1 0 100-2H7zM4 7a1 1 0 011-1h10a1 1 0 110 2H5a1 1 0 01-1-1zM2 11a2 2 0 012-2h12a2 2 0 012 2v4a2 2 0 01-2 2H4a2 2 0 01-2-2v-4z"></path>
+                  </svg>
+                  <span className="ml-3">Components</span>
+                </a>
+              </li>
+              <li>
+                <a
+                  href="#"
+                  className="flex items-center p-2 text-base font-normal text-gray-900 rounded-lg transition duration-75 hover:bg-gray-100 dark:hover:bg-gray-700 dark:text-white group"
+                >
+                  <svg
+                    aria-hidden="true"
+                    className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white"
                     fill="currentColor"
                     viewBox="0 0 20 20"
                     xmlns="http://www.w3.org/2000/svg"
                   >
                     <path
                       fillRule="evenodd"
+                      d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-2 0c0 .993-.241 1.929-.668 2.754l-1.524-1.525a3.997 3.997 0 00.078-2.183l1.562-1.562C15.802 8.249 16 9.1 16 10zm-5.165 3.913l1.58 1.58A5.98 5.98 0 0110 16a5.976 5.976 0 01-2.516-.552l1.562-1.562a4.006 4.006 0 001.789.027zm-4.677-2.796a4.002 4.002 0 01-.041-2.08l-.08.08-1.53-1.533A5.98 5.98 0 004 10c0 .954.223 1.856.619 2.657l1.54-1.54zm1.088-6.45A5.974 5.974 0 0110 4c.954 0 1.856.223 2.657.619l-1.54 1.54a4.002 4.002 0 00-2.346.033L7.246 4.668zM12 10a2 2 0 11-4 0 2 2 0 014 0z"
                       clipRule="evenodd"
-                      d="M8 4a4 4 0 100 8 4 4 0 000-8zM2 8a6 6 0 1110.89 3.476l4.817 4.817a1 1 0 01-1.414 1.414l-4.816-4.816A6 6 0 012 8z"
                     ></path>
                   </svg>
-                </div>
-                <input
-                  type="text"
-                  name="email"
-                  id="topbar-search"
-                  className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full pl-10 p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
-                  placeholder="Search"
-                />
-              </div>
-            </form>
-            <button
-          id="theme-toggle"
-          type="button"
-          className="text-gray-500 dark:text-gray-400 hover:bg-gray-100 dark:hover:bg-gray-700 focus:outline-none focus:ring-4 focus:ring-gray-200 dark:focus:ring-gray-700 rounded-lg text-sm p-2.5"
-        >
-          <svg
-            style={{ display: "block" }}
-            id="theme-toggle-dark-icon"
-            className="w-5 h-5"
-            fill="currentColor"
-            viewBox="0 0 20 20"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <path d="M17.293 13.293A8 8 0 016.707 2.707a8.001 8.001 0 1010.586 10.586z"></path>
-          </svg>
-          <svg
-            style={{ display: "none" }}
-            id="theme-toggle-light-icon"
-            className="w-5 h-5"
-            fill="currentColor"
-            viewBox="0 0 20 20"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <path
-              d="M10 2a1 1 0 011 1v1a1 1 0 11-2 0V3a1 1 0 011-1zm4 8a4 4 0 11-8 0 4 4 0 018 0zm-.464 4.95l.707.707a1 1 0 001.414-1.414l-.707-.707a1 1 0 00-1.414 1.414zm2.12-10.607a1 1 0 010 1.414l-.706.707a1 1 0 11-1.414-1.414l.707-.707a1 1 0 011.414 0zM17 11a1 1 0 100-2h-1a1 1 0 100 2h1zm-7 4a1 1 0 011 1v1a1 1 0 11-2 0v-1a1 1 0 011-1zM5.05 6.464A1 1 0 106.465 5.05l-.708-.707a1 1 0 00-1.414 1.414l.707.707zm1.414 8.486l-.707.707a1 1 0 01-1.414-1.414l.707-.707a1 1 0 011.414 1.414zM4 11a1 1 0 100-2H3a1 1 0 000 2h1z"
-              fillRule="evenodd"
-              clipRule="evenodd"
-            ></path>
-          </svg>
-        </button>
+                  <span className="ml-3">Help</span>
+                </a>
+              </li>
+            </ul>
           </div>
-          <div className="flex items-center lg:order-2">
-            <button
-              type="button"
-              className="flex mx-3 text-sm bg-gray-800 rounded-full md:mr-0 focus:ring-4 focus:ring-gray-300 dark:focus:ring-gray-600"
-              id="user-menu-button"
-              aria-expanded="false"
-              data-dropdown-toggle="dropdown"
+        </aside>
+
+        <main className={`md:${styles.mainBar} p-4 md:ml-72 h-auto pt-20`}>
+          <div className="border-gray-300 dark:border-gray-600 h-96 mb-4">
+
+            <ReactMarkdown
+              components={{
+                h1: CustomH1,
+                h2: CustomH2,
+                img: CustomImage,
+                hr: CustomDivider,
+              }}
             >
-              <span className="sr-only">Open user menu</span>
-            </button>
-            <div
-              className="hidden z-50 my-4 w-56 text-base list-none bg-white rounded divide-y divide-gray-100 shadow dark:bg-gray-700 dark:divide-gray-600 rounded-xl"
-              id="dropdown"
-            >
-              <div className="py-3 px-4">
-                <span className="block text-sm font-semibold text-gray-900 dark:text-white">
-                  Neil Sims
-                </span>
-                <span className="block text-sm text-gray-900 truncate dark:text-white">
-                  name@flowbite.com
-                </span>
-              </div>
-              <ul
-                className="py-1 text-gray-700 dark:text-gray-300"
-                aria-labelledby="dropdown"
-              >
-                <li>
-                  <a
-                    href="#"
-                    className="block py-2 px-4 text-sm hover:bg-gray-100 dark:hover:bg-gray-600 dark:text-gray-400 dark:hover:text-white"
-                  >
-                    My profile
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="block py-2 px-4 text-sm hover:bg-gray-100 dark:hover:bg-gray-600 dark:text-gray-400 dark:hover:text-white"
-                  >
-                    Account settings
-                  </a>
-                </li>
-              </ul>
-              <ul
-                className="py-1 text-gray-700 dark:text-gray-300"
-                aria-labelledby="dropdown"
-              >
-                <li>
-                  <a
-                    href="#"
-                    className="flex items-center py-2 px-4 text-sm hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white"
-                  >
-                    <svg
-                      className="mr-2 w-5 h-5 text-gray-400"
-                      fill="currentColor"
-                      viewBox="0 0 20 20"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        fillRule="evenodd"
-                        d="M3.172 5.172a4 4 0 015.656 0L10 6.343l1.172-1.171a4 4 0 115.656 5.656L10 17.657l-6.828-6.829a4 4 0 010-5.656z"
-                        clipRule="evenodd"
-                      ></path>
-                    </svg>
-                    My likes
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="flex items-center py-2 px-4 text-sm hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white"
-                  >
-                    <svg
-                      className="mr-2 w-5 h-5 text-gray-400"
-                      fill="currentColor"
-                      viewBox="0 0 20 20"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path d="M7 3a1 1 0 000 2h6a1 1 0 100-2H7zM4 7a1 1 0 011-1h10a1 1 0 110 2H5a1 1 0 01-1-1zM2 11a2 2 0 012-2h12a2 2 0 012 2v4a2 2 0 01-2 2H4a2 2 0 01-2-2v-4z"></path>
-                    </svg>
-                    Collections
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="flex justify-between items-center py-2 px-4 text-sm hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white"
-                  >
-                    <span className="flex items-center">
-                      <svg
-                        aria-hidden="true"
-                        className="mr-2 w-5 h-5 text-primary-600 dark:text-primary-500"
-                        fill="currentColor"
-                        viewBox="0 0 20 20"
-                        xmlns="http://www.w3.org/2000/svg"
-                      >
-                        <path
-                          fillRule="evenodd"
-                          d="M12.395 2.553a1 1 0 00-1.45-.385c-.345.23-.614.558-.822.88-.214.33-.403.713-.57 1.116-.334.804-.614 1.768-.84 2.734a31.365 31.365 0 00-.613 3.58 2.64 2.64 0 01-.945-1.067c-.328-.68-.398-1.534-.398-2.654A1 1 0 005.05 6.05 6.981 6.981 0 003 11a7 7 0 1011.95-4.95c-.592-.591-.98-.985-1.348-1.467-.363-.476-.724-1.063-1.207-2.03zM12.12 15.12A3 3 0 017 13s.879.5 2.5.5c0-1 .5-4 1.25-4.5.5 1 .786 1.293 1.371 1.879A2.99 2.99 0 0113 13a2.99 2.99 0 01-.879 2.121z"
-                          clipRule="evenodd"
-                        ></path>
-                      </svg>
-                      Pro version
-                    </span>
-                    <svg
-                      aria-hidden="true"
-                      className="w-5 h-5 text-gray-400"
-                      fill="currentColor"
-                      viewBox="0 0 20 20"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        fillRule="evenodd"
-                        d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
-                        clipRule="evenodd"
-                      ></path>
-                    </svg>
-                  </a>
-                </li>
-              </ul>
-              <ul
-                className="py-1 text-gray-700 dark:text-gray-300"
-                aria-labelledby="dropdown"
-              >
-                <li>
-                  <a
-                    href="#"
-                    className="block py-2 px-4 text-sm hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white"
-                  >
-                    Sign out
-                  </a>
-                </li>
-              </ul>
-            </div>
+              {markdownBody}
+            </ReactMarkdown>
           </div>
-        </div>
-      </nav>
-
-      <aside
-        id="default-sidebar"
-        className={`${styles.sidebar} fixed top-0 left-0 z-40 w-64 h-screen transition-transform -translate-x-full sm:translate-x-0`}
-        aria-label="Sidenav"
-      >
-        <div className="overflow-y-auto py-5 px-3 h-full bg-white border-r border-gray-200 dark:bg-gray-800 dark:border-gray-700">
-          <ul className="space-y-2">
-            {sidebarTabs.map((item : any) => (
-              <SidebarDocsTab key={item} item={item} />
-            ))}
-          </ul>
-
-          <ul className="pt-5 mt-5 space-y-2 border-t border-gray-200 dark:border-gray-700">
-            <li>
-              <a
-                href="#"
-                className="flex items-center p-2 text-base font-normal text-gray-900 rounded-lg transition duration-75 hover:bg-gray-100 dark:hover:bg-gray-700 dark:text-white group"
-              >
-                <svg
-                  aria-hidden="true"
-                  className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path d="M9 2a1 1 0 000 2h2a1 1 0 100-2H9z"></path>
-                  <path
-                    fillRule="evenodd"
-                    d="M4 5a2 2 0 012-2 3 3 0 003 3h2a3 3 0 003-3 2 2 0 012 2v11a2 2 0 01-2 2H6a2 2 0 01-2-2V5zm3 4a1 1 0 000 2h.01a1 1 0 100-2H7zm3 0a1 1 0 000 2h3a1 1 0 100-2h-3zm-3 4a1 1 0 100 2h.01a1 1 0 100-2H7zm3 0a1 1 0 100 2h3a1 1 0 100-2h-3z"
-                    clipRule="evenodd"
-                  ></path>
-                </svg>
-                <span className="ml-3">Docs</span>
-              </a>
-            </li>
-            <li>
-              <a
-                href="#"
-                className="flex items-center p-2 text-base font-normal text-gray-900 rounded-lg transition duration-75 hover:bg-gray-100 dark:hover:bg-gray-700 dark:text-white group"
-              >
-                <svg
-                  aria-hidden="true"
-                  className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path d="M7 3a1 1 0 000 2h6a1 1 0 100-2H7zM4 7a1 1 0 011-1h10a1 1 0 110 2H5a1 1 0 01-1-1zM2 11a2 2 0 012-2h12a2 2 0 012 2v4a2 2 0 01-2 2H4a2 2 0 01-2-2v-4z"></path>
-                </svg>
-                <span className="ml-3">Components</span>
-              </a>
-            </li>
-            <li>
-              <a
-                href="#"
-                className="flex items-center p-2 text-base font-normal text-gray-900 rounded-lg transition duration-75 hover:bg-gray-100 dark:hover:bg-gray-700 dark:text-white group"
-              >
-                <svg
-                  aria-hidden="true"
-                  className="flex-shrink-0 w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    fillRule="evenodd"
-                    d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-2 0c0 .993-.241 1.929-.668 2.754l-1.524-1.525a3.997 3.997 0 00.078-2.183l1.562-1.562C15.802 8.249 16 9.1 16 10zm-5.165 3.913l1.58 1.58A5.98 5.98 0 0110 16a5.976 5.976 0 01-2.516-.552l1.562-1.562a4.006 4.006 0 001.789.027zm-4.677-2.796a4.002 4.002 0 01-.041-2.08l-.08.08-1.53-1.533A5.98 5.98 0 004 10c0 .954.223 1.856.619 2.657l1.54-1.54zm1.088-6.45A5.974 5.974 0 0110 4c.954 0 1.856.223 2.657.619l-1.54 1.54a4.002 4.002 0 00-2.346.033L7.246 4.668zM12 10a2 2 0 11-4 0 2 2 0 014 0z"
-                    clipRule="evenodd"
-                  ></path>
-                </svg>
-                <span className="ml-3">Help</span>
-              </a>
-            </li>
-          </ul>
-        </div>
-      </aside>
-
-      <main className="p-4 md:ml-64 h-auto pt-20">
-        {/* <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4 mb-4">
-        <div
-          className="border-2 border-dashed border-gray-300 rounded-lg dark:border-gray-600 h-32 md:h-64"
-        ></div>
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-32 md:h-64"
-        ></div>
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-32 md:h-64"
-        ></div>
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-32 md:h-64"
-        ></div>
-      </div> */}
-        <div className="border-gray-300 dark:border-gray-600 h-96 mb-4">
-          {/* <MDXProvider components={components}> */}
-          {/* <Component />
-        </MDXProvider> */}
-          <ReactMarkdown
-            components={{
-              h1: CustomH1,
-              h2: CustomH2,
-            }}
-          >
-            {markdownBody}
-          </ReactMarkdown>
-        </div>
-        {/* <div className="grid grid-cols-2 gap-4 mb-4">
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72"
-        ></div>
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72"
-        ></div>
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72"
-        ></div>
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72"
-        ></div>
+        </main>
       </div>
-      <div
-        className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-96 mb-4"
-      ></div>
-      <div className="grid grid-cols-2 gap-4">
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72"
-        ></div>
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72"
-        ></div>
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72"
-        ></div>
-        <div
-          className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72"
-        ></div>
-      </div> */}
-      </main>
-    </div>
-  );
-}
+    );
+  }
+
