+++ app/docs/components/user/CustomLayout.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const CustomDivider = (props: any) => {
+  return (
+    <hr className="my-4" />
+  );
+};
+
+export { CustomDivider };
